App 3/requirements.txt
streamlit==1.31.1
gtts==2.5.1
SpeechRecognition==3.10.1
opencv-python==4.9.0.80
playsound==1.3.0
googletrans==3.1.0a0
pillow==10.2.0
numpy==1.26.4 